First pass at turning off auto layout for the grid content view. Cell subviews are not being laid out properly...?

+++ DMGridStackView.h
@@ -11,7 +11,7 @@
 @interface DMGridStackView : NSStackView
 
 @property (nonatomic, assign) BOOL visible;
-@property (nonatomic, assign) unsigned int column;
+@property (nonatomic, assign) NSUInteger column;
 
 @property (nonatomic, assign) NSRange rowsInStack;
